ajuste no fluxo de pessoas

+++ packages/pessoas/lib/data/remote/pontos_remote_data_source.dart
@@ -23,7 +23,7 @@ class PontosRemoteDataSource extends RemoteDataSourceBase
 
   @override
   Future<Ponto> postPonto({
-    required int valor,
+    required double valor,
     required DateTime validade,
     required String descricao,
     required int idPessoa,
