+++ packages/pessoas/lib/domain/data/data_sourcers/remote/i_pontos_remote_data_source.dart
@@ -2,7 +2,7 @@ import 'package:pessoas/models.dart';
 
 abstract class IPontosRemoteDataSource {
   Future<Ponto> postPonto({
-    required int valor,
+    required double valor,
     required DateTime validade,
     required String descricao,
     required int idPessoa,
